Train at half resolution of test

+++ bird_detector_paper/training_loop.py
@@ -227,7 +227,7 @@ def run(patch_size=2000):
     
     prepare_test(patch_size=patch_size)
     #Only open training raster once because its so huge
-    prepare_train(patch_size=patch_size/2)
+    prepare_train(patch_size=int(patch_size/2))
     
     p , r = training(proportion=0, pretrained=True, patch_size=patch_size)
